Rename missing alphabets to missing alphabets_c
write a program in c depending on the following problem
Given a sentence Str. The task is to find whether the given sentence contains all letters of the english alphabet (a to z or A to Z). If it does not, then print all the missing letters of the alphabet, otherwise print 0.
Note:
.All characters in the given sentence should be treated as case insensitive , which means  that 'A' and 'a' are to be treated as the same.
.The output should be in alphabetic order.
.The output should contain only lowercase alphabets.
.If none of the letters are missing print 0 as output.
Example 1:
Input:
The four boxing wizard starts over the quickly --- value of Str
Output:
jmp-- Missing letters
The program reads a sentence from the user, initializes an array of 26 elements (one for each letter of the alphabet), and iterates over the sentence to mark the letters that are present in the sentence. The program treats all letters as lowercase and ignores non-alphabetic characters.

After that, the program iterates over the alphabet array to find the missing letters. If a letter is missing, it is printed. If none of the letters are missing, the program prints "0".